+++ src/modern-async.mjs
@@ -15,8 +15,6 @@ export { default as findIndex } from './findIndex.mjs'
 export { default as forEach } from './forEach.mjs'
 export { default as map } from './map.mjs'
 export { default as mapGenerator } from './mapGenerator.mjs'
-export { default as mapLimit } from './mapLimit.mjs'
-export { default as mapSeries } from './mapSeries.mjs'
 export { default as Queue } from './Queue.mjs'
 export { default as queueMicrotask } from './queueMicrotask.mjs'
 export { default as reduce } from './reduce.mjs'
